split anatomy.py into multiple python files
