+++ client/ayon_core/pipeline/anatomy/__init__.py
@@ -1,0 +1,17 @@
+from .exceptions import (
+    ProjectNotSet,
+    RootCombinationError,
+    TemplateMissingKey,
+    AnatomyTemplateUnsolved,
+)
+from .anatomy import Anatomy
+
+
+__all__ = (
+    "ProjectNotSet",
+    "RootCombinationError",
+    "TemplateMissingKey",
+    "AnatomyTemplateUnsolved",
+
+    "Anatomy",
+)
